Refactor Utils to DL Utils under util directory
