+++ ingestion/tests/unit/topology/database/test_datalake.py
@@ -31,10 +31,7 @@ from metadata.generated.schema.metadataIngestion.workflow import (
 )
 from metadata.generated.schema.type.entityReference import EntityReference
 from metadata.ingestion.source.database.datalake.metadata import DatalakeSource
-from metadata.ingestion.source.database.datalake.utils import (
-    read_from_avro,
-    read_from_json,
-)
+from metadata.utils.datalake_utils import read_from_avro, read_from_json
 
 mock_datalake_config = {
     "source": {
